# Trivia-Nexus/Trivia-Nexus/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Quizara API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all main categories
      parameters:
        - name: parentId
          in: query
          required: false
          schema:
            type: string
          description: Filter by parent category ID (to get sub-categories)
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
  /categories/{categoryId}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a category by ID
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Category details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          description: Category not found
  /categories/{categoryId}/subcategories:
    get:
      operationId: listSubcategories
      tags: [categories]
      summary: List sub-categories for a given main category
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of sub-categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
  /categories/{categoryId}/stats:
    get:
      operationId: getCategoryStats
      tags: [categories]
      summary: Get question count and avg difficulty for a category
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Category stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryStats'
...
  /leaderboard:
    get:
      operationId: getLeaderboard
      tags: [leaderboard]
      summary: Get global leaderboard (individual)
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [all_time, weekly]
          description: Leaderboard time period
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardEntry'
...
  /leaderboard/me:
    get:
      operationId: getMyRank
      tags: [leaderboard]
      summary: Get current user's rank and stats
      responses:
        '200':
          description: Current user rank info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRankInfo'
        '401':
          description: Unauthorized
  /leaderboard/arena/me:
    get:
      operationId: getMyArenaStats
      tags: [leaderboard]
      summary: Get current user's arena stats and rank
      responses:
        '200':
          description: Current user arena stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyArenaStats'
        '401':
          description: Unauthorized
  /leaderboard/arena:
    get:
      operationId: getArenaLeaderboard
      tags: [leaderboard]
      summary: Get arena leaderboard ranked by wins then win-rate
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Arena leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArenaLeaderboardEntry'

  # Admin endpoints
...
components:
...
  schemas:
...
    Category:
      type: object
      required: [id, name, nameAr, icon, parentId, questionCount, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        nameAr:
          type: string
        icon:
          type: string
        color:
          type: [string, 'null']
        imageUrl:
          type: [string, 'null']
        parentId:
          type: [string, 'null']
        questionCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    CategoryStats:
      type: object
      required: [categoryId, questionCount, avgDifficulty, timesPlayed]
      properties:
        categoryId:
          type: string
        questionCount:
          type: integer
        avgDifficulty:
          type: number
        timesPlayed:
          type: integer
...
    LeaderboardEntry:
      type: object
      required: [rank, userId, username, totalScore, gamesPlayed, country]
      properties:
        rank:
          type: integer
        userId:
          type: string
        username:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        totalScore:
          type: integer
        gamesPlayed:
          type: integer
        country:
          type: [string, 'null']
...
    UserRankInfo:
      type: object
      required: [rank, totalScore, gamesPlayed, percentile]
      properties:
        rank:
          type: [integer, 'null']
        totalScore:
          type: integer
        gamesPlayed:
          type: integer
        percentile:
          type: number
...
    MyArenaStats:
      type: object
      required: [rank, wins, losses, draws, totalGames, totalScore, winRate]
      properties:
        rank:
          type: integer
          nullable: true
        wins:
          type: integer
        losses:
          type: integer
        draws:
          type: integer
        totalGames:
          type: integer
        totalScore:
          type: integer
        winRate:
          type: number
          format: float
    ArenaLeaderboardEntry:
      type: object
      required: [rank, userId, username, wins, losses, draws, totalGames, totalScore, winRate]
      properties:
        rank:
          type: integer
        userId:
          type: string
        username:
          type: string
          nullable: true
        profileImageUrl:
          type: string
          nullable: true
        wins:
          type: integer
        losses:
          type: integer
        draws:
          type: integer
        totalGames:
          type: integer
        totalScore:
          type: integer
        winRate:
          type: number
          format: float